review: split field_ref in file
