+++ query-engine/schema-builder/src/input_types/fields/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod arguments;
 pub(crate) mod data_input_mapper;
 pub(crate) mod field_filter_types;
+pub(crate) mod field_ref_type;
 pub(crate) mod input_fields;
 
 use super::*;
